Fixing integration
