+++ requirements.txt
@@ -7,3 +7,4 @@ wheel
 setuptools
 build
 twine
+auto_gpt_plugin_template
